LTD Co Contractor Company Checks
ed.  Then the employee can only be upgraded to "Processed and checked"
if nothing is ticked then it wont upgrade.   also then have this show on
the monthly branch managers health check report - have it so that it
will show a ltd company contractors details once its been 12 month since
the date it was checked.  have the report show the  First name and
surname of the worker then the LTD company  name, then the
co-registration number.